+++ ci4/app/Controllers/DeploymentSpecifications.php
@@ -482,6 +482,7 @@ class DeploymentSpecifications extends ResourceController {
                     $data->reclaim_policy,
                     $data->nfs_server,
                     $data->nfs_path,
+                    $data->storage_class,
                 ),
                 $body->values
             );